% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{Title A function that returns the dimensions and name of a data frame}
\usage{
get_info_data(data_frame)
}
\arguments{
\item{data_frame}{Function takes a data frame}
}
\value{
A list  of 2 dimensions and names of data frame
}
\description{
Description Given a data frame it returms a list of the names of the df and dimensions
}
\examples{
get_info_data(iris)
}
