% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{Title Functions gives the mean of numerics columns of a df}
\usage{
get_mean_data(data_frame)
}
\arguments{
\item{data_frame}{A data.frame}
}
\value{
a data frame / message indicating whether there was numeric cols found
}
\description{
Description
}
\examples{
get_mean_data(iris)

char_df = data.frame(A = c("A", "B"),
                     B = c("D", "B"))


get_mean_data(char_df)
}
